ok got the suite refactored into a subdirctoy. Need to add it to the CI and polish up our documentation.
