+++ integration_tests/sdk/requirements_tests/requirements_test.py
@@ -2,14 +2,15 @@ import pandas as pd
 import pytest
 from aqueduct.error import AqueductError
 from aqueduct import infer_requirements
-from constants import SENTIMENT_SQL_QUERY
-from requirements.model import sentiment_prediction_using_transformers
 from utils import get_integration_name
+
+from constants import SENTIMENT_SQL_QUERY
+from transformers_model.model import sentiment_prediction_using_transformers
 
 from aqueduct import op
 
 INVALID_REQUIREMENTS_PATH = "~/random.txt"
-VALID_REQUIREMENTS_PATH = "requirements/requirements.txt"
+VALID_REQUIREMENTS_PATH = "transformers_model/requirements.txt"
 
 
 def _transformers_package_exists():
@@ -53,7 +54,6 @@ def _check_infer_requirements(transformers_exists: bool):
     else:
         assert all(not req_str.startswith("transformers") for req_str in inferred_reqs)
 
-
 @op()
 def sentiment_prediction_without_reqs_path(reviews: pd.DataFrame) -> pd.DataFrame:
     import transformers
@@ -61,6 +61,7 @@ def sentiment_prediction_without_reqs_path(reviews: pd.DataFrame) -> pd.DataFram
     return reviews.join(pd.DataFrame(model(list(reviews["review"]))))
 
 
+@pytest.mark.requirements
 def test_infer_requirements(client):
     if _transformers_package_exists():
         _uninstall_transformers_package()
@@ -87,8 +88,11 @@ def test_infer_requirements(client):
 def sentiment_prediction_with_invalid_reqs_path(table: pd.DataFrame) -> pd.DataFrame:
     return table
 
-
+@pytest.mark.requirements
 def test_invalid_requirements_path(client):
+    if _transformers_package_exists():
+        _uninstall_transformers_package()
+
     db = client.integration(name=get_integration_name())
     table = db.sql(query=SENTIMENT_SQL_QUERY)
 
@@ -99,13 +103,16 @@ def test_invalid_requirements_path(client):
 
 @op(requirements=VALID_REQUIREMENTS_PATH)
 def sentiment_prediction_with_valid_reqs_path(reviews: pd.DataFrame) -> pd.DataFrame:
-    """This uses the requirements.txt in the requirements/ subdirectory to install transformers."""
+    """This uses the requirements.txt in the transformers_model/ subdirectory to install transformers."""
     import transformers
     model = transformers.pipeline("sentiment-analysis")
     return reviews.join(pd.DataFrame(model(list(reviews["review"]))))
 
-
+@pytest.mark.requirements
 def test_requirements_installation_from_path(client):
+    if _transformers_package_exists():
+        _uninstall_transformers_package()
+
     db = client.integration(name=get_integration_name())
     table = db.sql(query=SENTIMENT_SQL_QUERY)
     valid_path_table = sentiment_prediction_with_valid_reqs_path(table)
@@ -119,19 +126,26 @@ def sentiment_prediction_with_string_requirements(reviews):
     return reviews.join(pd.DataFrame(model(list(reviews["review"]))))
 
 
+@pytest.mark.requirements
 def test_requirements_installation_from_strings(client):
     # TODO(test the list if not well-formed.)
+    if _transformers_package_exists():
+        _uninstall_transformers_package()
+
     db = client.integration(name=get_integration_name())
     table = db.sql(query=SENTIMENT_SQL_QUERY)
     valid_path_table = sentiment_prediction_with_string_requirements(table)
     assert valid_path_table.get().shape[0] == 100
 
-
+@pytest.mark.requirements
 def test_default_requirements_installation(client):
     """
-    This uses the decorated function in the requirements/ subdirectory, which already has a requirements.txt
+    This uses the decorated function in the transformers_model/ subdirectory, which already has a requirements.txt
     that it should be installing.
     """
+    if _transformers_package_exists():
+        _uninstall_transformers_package()
+
     db = client.integration(name=get_integration_name())
     table = db.sql(query=SENTIMENT_SQL_QUERY)
     valid_path_table = sentiment_prediction_using_transformers(table)
