+++ integration_tests/sdk/requirements_tests/transformers_model/model.py
