+++ integration_tests/sdk/requirements_tests/transformers_model/requirements.txt
